refactored post content
