+++ src/app/containers/post/postContent/PostContent.jsx
@@ -1,15 +1,12 @@
 import React, { Component } from 'react';
 import { Link, browserHistory } from 'react-router';
 import styled from 'styled-components';
-import { connect } from 'react-redux';
-import { createSelector } from 'reselect';
 
 import Icon from 'golos-ui/Icon';
 import { TagLink } from 'golos-ui/Tag';
 
 import PostHeader from 'src/app/containers/post/PostHeader';
 import MarkdownViewer from 'app/components/cards/MarkdownViewer';
-import { currentPostSelector } from 'src/app/redux/selectors/post/commonPost';
 
 const Wrapper = styled.section`
     position: relative;
@@ -59,7 +56,7 @@ const PostBody = styled.div`
     padding: 12px 0 14px;
 
     p {
-        color: #373D3F;
+        color: #373d3f;
         font-family: 'Open Sans', sans-serif;
         font-size: 18px;
         letter-spacing: -0.29px;
@@ -88,7 +85,7 @@ const CategoryWrapper = styled.div`
     justify-content: space-between;
 `;
 
-class PostContent extends Component {
+export class PostContent extends Component {
     onBackClick = e => {
         e.preventDefault();
         browserHistory.goBack();
@@ -155,33 +152,3 @@ class PostContent extends Component {
         );
     }
 }
-
-const mapStateToProps = createSelector(
-    [currentPostSelector, state => state.ui.location],
-    (post, location) => {
-        const prev = location.get('previous');
-
-        let backUrl = null;
-
-        if (prev) {
-            backUrl = prev.get('pathname') + prev.get('search') + prev.get('hash');
-        }
-
-        return {
-            tags: post.tags,
-            payout: post.payout,
-            data: post.data,
-            category: post.category,
-            title: post.title,
-            body: post.body,
-            jsonMetadata: post.jsonMetadata,
-            pictures: post.pictures,
-            created: post.created,
-            permLink: post.permLink,
-            isPromoted: post.promotedAmount > 0,
-            backUrl,
-        };
-    }
-);
-
-export default connect(mapStateToProps)(PostContent);
