+++ src/app/containers/post/postContent/index.js
@@ -1,0 +1,2 @@
+export { default } from './PostContent.connect';
+export { PostContent } from './PostContent';
